# API_Back-End/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Livros",
    "version": "1.0.0",
    "description": "Documentação da API para gerenciamento de livros"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/livros": {
      "get": {
        "summary": "Lista todos os livros",
        "tags": ["Livros"],
        "responses": {
          "200": {
            "description": "Lista de livros retornada com sucesso"
          }
        }
      },
      "post": {
        "summary": "Adiciona um novo livro",
        "tags": ["Livros"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["titulo", "autor", "ano", "preco", "foto"],
                "properties": {
                  "titulo": { "type": "string" },
                  "autor": { "type": "string" },
                  "ano": { "type": "integer" },
                  "preco": { "type": "number" },
                  "foto": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Livro criado com sucesso" },
          "400": { "description": "Erro nos dados enviados" }
        }
      }
    },
    "/livros/{id}": {
      "put": {
        "summary": "Atualiza um livro (todos os campos)",
        "tags": ["Livros"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do livro a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": { "type": "string" },
                  "autor": { "type": "string" },
                  "ano": { "type": "integer" },
                  "preco": { "type": "number" },
                  "foto": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Livro atualizado com sucesso" },
          "404": { "description": "Livro não encontrado" }
        }
      },
      "patch": {
        "summary": "Atualiza parcialmente um livro (apenas os campos informados)",
        "tags": ["Livros"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do livro a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": { "type": "string" },
                  "autor": { "type": "string" },
                  "ano": { "type": "integer" },
                  "preco": { "type": "number" },
                  "foto": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Livro atualizado parcialmente com sucesso" },
          "400": { "description": "Nenhum campo enviado para atualização" },
          "404": { "description": "Livro não encontrado" }
        }
      },
      "delete": {
        "summary": "Exclui um livro",
        "tags": ["Livros"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do livro a ser excluído"
          }
        ],
        "responses": {
          "200": { "description": "Livro deletado com sucesso" },
          "404": { "description": "Livro não encontrado" }
        }
      }
    },
    "/livros/{id}/status": {
      "patch": {
        "summary": "Atualiza o status de um livro (ativo ou inativo)",
        "tags": ["Livros"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do livro a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ativo": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Livro atualizado com sucesso" },
          "400": { "description": "Erro nos dados enviados" },
          "404": { "description": "Livro não encontrado" }
        }
      }
    },
    "/livros/filtro/{palavra}": {
      "get": {
        "summary": "Filtra livros pelo título ou autor",
        "tags": ["Livros"],
        "parameters": [
          {
            "in": "path",
            "name": "palavra",
            "required": true,
            "schema": { "type": "string" },
            "description": "Palavra-chave para busca no título ou autor"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de livros filtrada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "titulo": { "type": "string" },
                      "autor": { "type": "string" },
                      "ano": { "type": "integer" },
                      "preco": { "type": "number" },
                      "foto": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Campo vazio ou palavra inválida." },
          "500": { "description": "Erro interno no servidor." }
        }
      }
    },
    "/livros/dados/resumo": {
      "get": {
        "summary": "Obtém um resumo dos livros",
        "tags": ["Dados"],
        "responses": {
          "200": { "description": "Resumo retornado com sucesso" },
          "404": { "description": "Nenhum dado encontrado" }
        }
      }
    },
    "/livros/dados/grafico": {
      "get": {
        "summary": "Obtém dados para gráfico",
        "tags": ["Dados"],
        "responses": {
          "200": { "description": "Dados retornados com sucesso" },
          "404": { "description": "Nenhum dado encontrado" }
        }
      }
    },
    "/livros/dados/registros/{ano}": {
      "get": {
        "summary": "Obtém os registros de livros para um ano específico",
        "tags": ["Dados"],
        "parameters": [
          {
            "in": "path",
            "name": "ano",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Ano para o qual os registros de livros serão retornados"
          }
        ],
        "responses": {
          "200": {
            "description": "Registros de livros para o ano retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "titulo": { "type": "string" },
                      "autor": { "type": "string" },
                      "ano": { "type": "integer" },
                      "preco": { "type": "number" },
                      "foto": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhum dado encontrado para o ano solicitado"
          }
        }
      }
    }
  }
}